Extract MemberPage to its own file
This step is required so the scraper_test can run tests against it in a future
commit.
