+++ scraper.rb
@@ -10,6 +10,8 @@ require 'open-uri/cached'
 OpenURI::Cache.cache_path = '.cache'
 # require 'scraped_page_archive/open-uri'
 
+require_rel 'lib'
+
 def scrape(h)
   url, klass = h.to_a.first
   klass.new(response: Scraped::Request.new(url: url).response)
@@ -54,36 +56,6 @@ class WinnerRow < Scraped::HTML
 
   def tds
     noko.css('td')
-  end
-end
-
-class MemberPage < Scraped::HTML
-  decorator Scraped::Response::Decorator::AbsoluteUrls
-
-  field :id do
-    File.basename(url, '.*')
-  end
-
-  field :name do
-    name_td.text.delete('*')
-  end
-
-  field :image do
-    name_td.xpath('preceding::img[1]/@src').text
-  end
-
-  field :party_id do
-    party_node.text[/\(([^)]+)\)/, 1]
-  end
-
-  private
-
-  def name_td
-    noko.css('.Article02')
-  end
-
-  def party_node
-    name_td.xpath('following::td[.="Party"]//following-sibling::td//a[contains(@href, "/parties/")]')
   end
 end
